+++ robocorp-code/vscode-client/src/robocorpSettings.ts
@@ -18,6 +18,7 @@ export const ROBOCORP_RCC_CONFIG_LOCATION = "robocorp.rcc.config_location";
 export const ROBOCORP_HOME = "robocorp.home";
 export const ROBOCORP_VERIFY_LSP = "robocorp.verifyLSP";
 export const ROBOCORP_AUTO_SET_PYTHON_EXTENSION_INTERPRETER = "robocorp.autoSetPythonExtensionInterpreter";
+export const ROBOCORP_AUTO_SET_PYTHON_EXTENSION_DISABLE_ACTIVATE_TERMINAL = "robocorp.autoSetPythonExtensionDisableActivateTerminal";
 
 export function getLanguageServerTcpPort(): number {
     let key = ROBOCORP_LANGUAGE_SERVER_TCP_PORT;
@@ -147,6 +148,21 @@ export function getAutosetpythonextensioninterpreter(): boolean {
 
 export async function setAutosetpythonextensioninterpreter(value): Promise<void> {
     let key = ROBOCORP_AUTO_SET_PYTHON_EXTENSION_INTERPRETER;
+    let i = key.lastIndexOf('.');
+
+    let config = workspace.getConfiguration(key.slice(0, i));
+    await config.update(key.slice(i + 1), value, ConfigurationTarget.Global);
+}
+
+
+export function getAutosetpythonextensiondisableactivateterminal(): boolean {
+    let key = ROBOCORP_AUTO_SET_PYTHON_EXTENSION_DISABLE_ACTIVATE_TERMINAL;
+    return get<boolean>(key);
+}
+
+
+export async function setAutosetpythonextensiondisableactivateterminal(value): Promise<void> {
+    let key = ROBOCORP_AUTO_SET_PYTHON_EXTENSION_DISABLE_ACTIVATE_TERMINAL;
     let i = key.lastIndexOf('.');
 
     let config = workspace.getConfiguration(key.slice(0, i));