+++ robocorp-code/docs/changelog.md
@@ -1,3 +1,13 @@
+NEXT
+-----------------------------
+
+- Automatically set `"python.terminal.activateEnvironment": false`. [#636](https://github.com/robocorp/robotframework-lsp/issues/636)
+  - This fixes the case where the `ms-python` extension tries to activate an environment when starting a Robot terminal.
+  - It's possible to disable this by setting `"robocorp.autoSetPythonExtensionDisableActivateTerminal": false`.
+- `Set pythonPath based on robot.yaml` was renamed to `Set python executable based on robot.yaml`.
+- When setting the default interpreter to match the python from the robot, only the internal API is used as `python.pythonPath` is now deprecetad.
+- When launching the python debugger, the python executable is properly set in the generated launch configuration with `python` instead of `pythonPath`.
+
 New in 0.28.0 (2022-03-14)
 -----------------------------